+++ plain-color.c
@@ -2,7 +2,7 @@
 #include "plain-color.h"
 #include "ipoint.h"
 
-void set_plain_colors(struct shape *s, float r, float g, float b)
+void set_plain_colors(struct shape *s, double r, double g, double b)
 {
 	if (!s->paint_data)
 		s->paint_data = malloc(sizeof(struct plain_color));
@@ -13,7 +13,7 @@ void set_plain_colors(struct shape *s, float r, float g, float b)
 }
 
 
-void set_plain_color(struct shape *s, int side, float r, float g, float b)
+void set_plain_color(struct shape *s, int side, double r, double g, double b)
 {
 	int c = 0;
