+++ rgb.h
@@ -2,26 +2,26 @@
 #define RGB_H
 
 struct rgb {
-	float r;
-	float g;
-	float b;
+	double r;
+	double g;
+	double b;
 };
 
-#define RGBWhite	1.00F, 1.00F, 1.00F
-#define RGBBlack	0.00F, 0.00F, 0.00F
+#define RGBWhite	1.00, 1.00, 1.00
+#define RGBBlack	0.00, 0.00, 0.00
 
-#define RGBRed		1.00F, 0.00F, 0.00F
-#define RGBGreen	0.00F, 1.00F, 0.00F
-#define RGBBlue		0.00F, 0.00F, 1.00F
+#define RGBRed		1.00, 0.00, 0.00
+#define RGBGreen	0.00, 1.00, 0.00
+#define RGBBlue		0.00, 0.00, 1.00
 
-#define RGBYellow	1.00F, 1.00F, 0.00F
-#define RGBMagenta	1.00F, 0.00F, 1.00F
-#define RGBCyan		0.00F, 1.00F, 1.00F
+#define RGBYellow	1.00, 1.00, 0.00
+#define RGBMagenta	1.00, 0.00, 1.00
+#define RGBCyan		0.00, 1.00, 1.00
 
-#define RGBLightGray	0.80F, 0.80F, 0.80F
-#define RGBDarkGray	0.20F, 0.20F, 0.20F
-#define RGBLightBlue	0.06F, 0.06F, 0.86F
-#define RGBDarkBlue	0.06F, 0.06F, 0.33F
+#define RGBLightGray	0.80, 0.80, 0.80
+#define RGBDarkGray	0.20, 0.20, 0.20
+#define RGBLightBlue	0.06, 0.06, 0.86
+#define RGBDarkBlue	0.06, 0.06, 0.33
 
 #define White		{ RGBWhite	 }
 #define Black		{ RGBBlack	 }
