+++ surfaces.h
@@ -4,12 +4,12 @@
 #include "shape.h"
 
 extern struct shape *plane(const struct coord*, const struct coord*);
-extern struct shape *sphere(const struct coord*, float);
+extern struct shape *sphere(const struct coord*, double);
 extern struct shape *paraboloid(const struct coord*, const struct coord*,
-				float, float, float, float);
+				double, double, double, double);
 extern struct shape *cylinder(const struct coord*, const struct coord*,
-			      float, float);
+			      double, double);
 extern struct shape *cone(const struct coord*, const struct coord*,
-			  float, float);
+			  double, double);
 
 #endif /* SURFACES_H */
